# src/openapi.yaml
openapi: '3.0.0'
info:
  version: 1.0.0
  title: Cache API
  description: A simple Cache API
  termsOfService: http://swagger.io/terms/
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /api
components:
  schemas:
    CacheEntry:
      properties:
        id:
          type: string
        key:
          type: string
        value:
          type: string
        expiresAt:
          type: string
          format: date-time
      required:
        - id
        - key
        - value
        - expiresAt
    PaginatedCacheEntry:
      properties:
        data:
          type: object
          properties:
            entries:
              type: array
              default: []
              items:
                $ref: '#/components/schemas/CacheEntry'

            next:
              type: string
          required:
            - entries
      required:
        - data
    Error:
      properties:
        message:
          type: string
      required:
        - message

paths:
  /cache:
    get:
      x-eov-operation-handler: cacheEntry.controller
      description: |
        List Cache entries
      operationId: listCacheEntries
      parameters:
        - name: limit
          in: query
          schema:
            type: number
          required: false
        - name: after
          in: query
          schema:
            type: string
          required: false
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCacheEntry'

        default:
          description: Unexpected Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      x-eov-operation-handler: cacheEntry.controller
      description: |
        Add or update a cache entry by key
      operationId: addOrUpdateCacheEntry
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                key:
                  type: string
                value:
                  type: string
                ttl:
                  type: number
              required:
                - key
                - value
        required: true
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheEntry'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /cache/{id}:
    get:
      x-eov-operation-handler: cacheEntry.controller
      description: |
        Get Cache entry by key
      operationId: getCacheEntryByKey
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                required:
                  - data
        # Does this match requirements? since we're generating dummy data
        '404':
          description: NOT_FOUND
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: string
                    nullable: true
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'